modules/bridge/pyproject.toml
[project]
name = "retailsense-bridge"
version = "1.0.0"
description = "MQTT to Kafka bridge for RetailSense"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "kafka-python-ng>=2.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/retailsense_bridge"]

[tool.hatch.build.targets.sdist]
include = ["src/retailsense_bridge", "tests"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true